Add redirection
At the DataLad handbook, we use this "hack" to distribute stable URLs.
Instead of relying on unchanged file names and directories, this allows
to reference a page via its reference handle (e.g., .. _openNeuro:) as a
redirection. Any key in :redirect:`value` can be added to a URL following
brainhack-princeton.github.io/handbook/r.html?value, and the URL will
redirect to the page associated with :ref:`value`.

We use this to safeguard important pages that are linked somewhere else
against file name changes or restructuring.